+++ server/utils/AiProviders/openRouter/index.js
@@ -7,7 +7,7 @@ const {
 const fs = require("fs");
 const path = require("path");
 const { safeJsonParse } = require("../../http");
-const logger = require("../../logging");
+const logger = require("../../logger");
 const cacheFolder = path.resolve(
   process.env.STORAGE_DIR
     ? path.resolve(process.env.STORAGE_DIR, "models", "openrouter")
