+++ server/utils/vectorDbProviders/pinecone/index.js
@@ -5,7 +5,7 @@ const { storeVectorResult, cachedVectorInformation } = require("../../files");
 const { v4: uuidv4 } = require("uuid");
 const { toChunks, getEmbeddingEngineSelection } = require("../../helpers");
 const { sourceIdentifier } = require("../../chats");
-const logger = require("../../../utils/logging");
+const logger = require("../../logger");
 
 const PineconeDB = {
   name: "Pinecone",
