+++ server/utils/vectorDbProviders/zilliz/index.js
@@ -10,7 +10,7 @@ const { v4: uuidv4 } = require("uuid");
 const { storeVectorResult, cachedVectorInformation } = require("../../files");
 const { toChunks, getEmbeddingEngineSelection } = require("../../helpers");
 const { sourceIdentifier } = require("../../chats");
-const logger = require("../../../utils/logging");
+const logger = require("../../logger");
 
 // Zilliz is basically a copy of Milvus DB class with a different constructor
 // to connect to the cloud
